requirements.txt
fastapi
uvicorn
pydantic
pandas
openpyxl
sqlite3  # Comes with Python stdlib
matplotlib
langchain
langchain-community
google-generativeai  # For Gemini
